+++ addon/gradients/gradient4.js
@@ -2,9 +2,9 @@
   @module ember-flexberry-gis
 */
 export default Object({
-  name: "grad4",
+  name: 'grad4',
 
-  colorS: "#6464c8",
+  colorS: '#6464c8',
 
-  colorE: "#ff9616"
+  colorE: '#ff9616'
 });