+++ addon/gradients/gradient5.js
@@ -2,9 +2,9 @@
   @module ember-flexberry-gis
 */
 export default Object({
-  name: "grad5",
+  name: 'grad5',
 
-  colorS: "#8e707d",
+  colorS: '#8e707d',
 
-  colorE: "#b992e0"
+  colorE: '#b992e0'
 });